TodoBot/pyproject.toml
[project]
name = "todobot"
version = "0.1.0"
description = "Telegram bot for managing Todo tasks via API FastAPI"
authors = [
    {name = "Admin"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-telegram-bot (>=22.3,<23.0)",
    "requests (>=2.32.5,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "cryptography (>=45.0.7,<46.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
